Move BaseEditor into Homepage builder
