+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Editor/Editor.tsx
