+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Editor/index.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 import { Box } from '@citizenlab/cl2-component-library';
 import { SerializedNodes } from '@craftjs/core';
 
-import BaseEditor from 'components/admin/ContentBuilder/Editor';
 import AccordionMultiloc from 'components/admin/ContentBuilder/Widgets/AccordionMultiloc';
 import ButtonMultiloc from 'components/admin/ContentBuilder/Widgets/ButtonMultiloc';
 import Container from 'components/admin/ContentBuilder/Widgets/Container';
@@ -16,6 +15,8 @@ import TwoColumn from 'components/admin/ContentBuilder/Widgets/TwoColumn';
 import WhiteSpace from 'components/admin/ContentBuilder/Widgets/WhiteSpace';
 
 import { WIDGETS } from '../Widgets';
+
+import BaseEditor from './Editor';
 
 type EditorProps = {
   children?: React.ReactNode;
