+++ front/app/containers/Admin/reporting/components/ReportBuilder/Editor/RenderNode/index.tsx
@@ -50,13 +50,10 @@ const RenderNode = ({ render }) => {
   } = useNode((node) => {
     // This can sometimes be undefined, even though
     // craftjs says it can't
-    // TODO: Fix this the next time the file is edited.
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     if (!node) return {};
-    // TODO: Fix this the next time the file is edited.
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     if (!node.data) return {};
-    // TODO: Fix this the next time the file is edited.
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     if (!node.events) return {};
